creada programa de control principal, separada api
