+++ edog_leg_trayectory.py
